move ASW specific types/helpers/constants out of common folder
