+++ frontend/src/features/permits/helpers/bridgeCalculationHelper.ts
