+++ frontend/src/features/permits/hooks/useRunAxleCalculation.ts
@@ -1,11 +1,13 @@
 import { Policy } from "onroute-policy-engine";
 import { Nullable } from "../../../common/types/common";
 import { PermitVehicleDetails } from "../types/PermitVehicleDetails";
-import { AxleConfiguration } from "../../../common/types/AxleUnit";
 import { PermitVehicleConfiguration } from "../types/PermitVehicleConfiguration";
 import { getDefaultVehicleConfiguration } from "../helpers/vehicles/configuration/getDefaultVehicleConfiguration";
 import { PermitType } from "../types/PermitType";
-import { VehicleConfiguration } from "onroute-policy-engine/types";
+import {
+  AxleConfiguration,
+  VehicleConfiguration,
+} from "onroute-policy-engine/types";
 
 export const useRunAxleCalculation = (policyEngine: Nullable<Policy>) => {
   return {
