+++ frontend/src/features/permits/types/AxleCalculationResult.ts
