+++ frontend/src/features/permits/types/BridgeCalculationResult.ts
